requirements.txt
fastapi 
uvicorn[standard]
pandas
numpy
python-multipart
dill
pyarrow
google-cloud-storage